Analysis of facilities provided on Austin-Bergstrom International Airport
In this report of analytical customer relationship management, it provides the data mining techniques that can tell the factors affecting the overall satisfaction rate of Austin-Bergstrom International Airport. The linear regression model is performed for the analysis. This analysis is done on the various facilities that the passengers can use by making use of analytical tool.